Revert "Update hyperkit go module"
The updated hyperkit go module seems to be the root cause of
https://github.com/code-ready/crc/issues/2158

I've confirmed that by just changing the version of the hyperkit module
the bug comes and go. crc/bundle/machine driver code are all unchanged
in these tests.
github.com/moby/hyperkit v0.0.0-20171206102831-721ca707b187 hangs after
a few minutes
github.com/moby/hyperkit v0.0.0-20171206102043-1d2506619d57 runs for 15+
minutes without hanging
I did not test the few commits in between

This is fairly mysterious at this point as the hyperkit commandline is
identical, qcow-tool is not installed so this should not related to the
new qcow-tool code in the updated module.

This reverts commit a0c6e877a4b17ea93d44d5abc9489faa029b208b.